2024-May/src/helper_functions.py
def square(num):
    # Checks if the input is a square number
    return num == int(num ** 0.5) ** 2


def palindrome(num):
    # Checks if the input is a palindrome number
    return str(num) == str(num)[::-1]


def one_more_than_palindrome(num):
    # Checks if the input is one more than a palindrome number
    return palindrome(num - 1)


def one_less_than_palindrome(num):
    # Checks if the input is one less than a palindrome number
    return palindrome(num + 1)


def prime_raised_to_prime_power(num):
    # Checks if the input is a prime number raised to a prime power
    if num < 2:
        return False
    for i in range(2, int(num ** 0.5) + 1):
        if num % i == 0:
            return False
    return prime(num)


def prime(num):
    # Checks if the input is a prime number
    if num < 2:
        return False
    for i in range(2, int(num ** 0.5) + 1):
        if num % i == 0:
            return False
    return True


def sum_of_digits_is_7(num):
    # Checks if the sum of the digits of the input is 7
    return sum(map(int, str(num))) == 7


def fibonacci(num):
    # Checks if the input is a Fibonacci number
    a, b = 0, 1
    while b < num:
        a, b = b, a + b
    return b == num


def multiple_of_37(num):
    # Checks if the input is a multiple of 37
    return num % 37 == 0


def palindrome_multiple_of_23(num):
    # Checks if the input is a palindrome number that is a multiple of 23
    return palindrome(num) and num % 23 == 0


def product_of_digits_ends_in_1(num):
    # Checks if the product of the digits of the input ends in 1
    product = 1
    for digit in str(num):
        product *= int(digit)
    return product % 10 == 1


def multiple_of_88(num):
    # Checks if the input is a multiple of 88
    return num % 88 == 0


def power_of_7(num):
    # Checks if the input is a power of 7
    if num < 1:
        return False
    while num % 7 == 0:
        num /= 7
    return num == 1


def multiple_of_5(num):
    # Checks if the input is a multiple of 5
    return num % 5 == 0


def cube(num):
    if num < 0:
        return -cube(-num)
        # Approximate cube root
    cube_root = round(num ** (1 / 3))
    return num == cube_root ** 3

